Create dedicated length_prefixed module
