+++ packages/storage/src/bucket.rs
@@ -5,9 +5,10 @@ use cosmwasm_std::{to_vec, ReadonlyStorage, StdResult, Storage};
 #[cfg(feature = "iterator")]
 use cosmwasm_std::{Order, KV};
 
+use crate::length_prefixed::{key_prefix, key_prefix_nested};
 #[cfg(feature = "iterator")]
 use crate::namespace_helpers::range_with_prefix;
-use crate::namespace_helpers::{get_with_prefix, key_prefix, key_prefix_nested, set_with_prefix};
+use crate::namespace_helpers::{get_with_prefix, set_with_prefix};
 #[cfg(feature = "iterator")]
 use crate::type_helpers::deserialize_kv;
 use crate::type_helpers::{may_deserialize, must_deserialize};
